+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `output_dir` parameter to `dataset_download` to download datasets or files to a custom location.
 * Add `output_dir` and `overwrite` parameters to `model_download` for direct model downloads.
 * Add `output_dir` and `overwrite` parameters to `notebook_output_download` for direct notebook output downloads.
+* Add `output_dir` and `overwrite` parameters to `competition_download` for direct competition downloads.
 
 ## v0.4.2 (January 29, 2026)